+++ app/src/main/java/com/cornellappdev/uplift/MainActivity.kt
@@ -7,12 +7,9 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.datastore.preferences.preferencesDataStore
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.cornellappdev.uplift.models.DatastoreRepository
+import com.cornellappdev.uplift.data.repositories.DatastoreRepository
 import com.cornellappdev.uplift.ui.MainNavigationWrapper
-import com.cornellappdev.uplift.ui.screens.ProfileCreationScreen
 import com.cornellappdev.uplift.ui.theme.UpliftTheme
-import com.cornellappdev.uplift.ui.viewmodels.ReportViewModel
 import com.cornellappdev.uplift.util.LockScreenOrientation
 import com.cornellappdev.uplift.util.PREFERENCES_NAME
 import dagger.hilt.android.AndroidEntryPoint
