+++ app/src/main/java/com/cornellappdev/uplift/data/models/Sport.kt
@@ -1,4 +1,4 @@
-package com.cornellappdev.uplift.models
+package com.cornellappdev.uplift.data.models
 
 class Sport(val painterId: Int, val name: String) : Comparable<Sport> {
     override fun compareTo(other: Sport): Int {
