+++ app/src/main/java/com/cornellappdev/uplift/data/repositories/DatastoreRepository.kt
@@ -1,4 +1,4 @@
-package com.cornellappdev.uplift.models
+package com.cornellappdev.uplift.data.repositories
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
