+++ app/src/main/java/com/cornellappdev/uplift/ui/MainNavigationWrapper.kt
@@ -19,18 +19,16 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
-import com.cornellappdev.uplift.nav.BottomNavScreens
-import com.cornellappdev.uplift.nav.popBackClass
-import com.cornellappdev.uplift.nav.popBackGym
-import com.cornellappdev.uplift.networking.ApiResponse
+import com.cornellappdev.uplift.ui.nav.BottomNavScreens
+import com.cornellappdev.uplift.ui.nav.popBackClass
+import com.cornellappdev.uplift.ui.nav.popBackGym
+import com.cornellappdev.uplift.data.models.ApiResponse
 import com.cornellappdev.uplift.ui.screens.ClassDetailScreen
 import com.cornellappdev.uplift.ui.screens.ClassScreen
 import com.cornellappdev.uplift.ui.screens.GymDetailScreen
