+++ app/src/main/java/com/cornellappdev/uplift/ui/components/gymdetail/GymCourtSection.kt
@@ -25,8 +25,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cornellappdev.uplift.R
-import com.cornellappdev.uplift.models.CourtFacility
-import com.cornellappdev.uplift.models.TimeInterval
+import com.cornellappdev.uplift.data.models.CourtFacility
+import com.cornellappdev.uplift.data.models.TimeInterval
 import com.cornellappdev.uplift.ui.components.gymdetail.TimeFlag
 import com.cornellappdev.uplift.util.ACCENT_CLOSED
 import com.cornellappdev.uplift.util.PRIMARY_BLACK
