+++ app/src/main/java/com/cornellappdev/uplift/ui/components/gymdetail/GymEquipmentGroup.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.style.TextAlign
 //import androidx.compose.ui.tooling.data.EmptyGroup.name
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.cornellappdev.uplift.models.EquipmentGrouping
+import com.cornellappdev.uplift.data.models.EquipmentGrouping
 import com.cornellappdev.uplift.util.PRIMARY_BLACK
 import com.cornellappdev.uplift.util.montserratFamily
