+++ app/src/main/java/com/cornellappdev/uplift/ui/components/gymdetail/GymFacilitySection.kt
@@ -32,8 +32,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cornellappdev.uplift.R
-import com.cornellappdev.uplift.models.OpenType
-import com.cornellappdev.uplift.models.UpliftGym
+import com.cornellappdev.uplift.data.models.OpenType
+import com.cornellappdev.uplift.data.models.UpliftGym
 import com.cornellappdev.uplift.ui.screens.LineSpacer
 import com.cornellappdev.uplift.util.ACCENT_CLOSED
 import com.cornellappdev.uplift.util.ACCENT_OPEN
