+++ app/src/main/java/com/cornellappdev/uplift/ui/components/gymdetail/GymHours.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cornellappdev.uplift.R
-import com.cornellappdev.uplift.models.TimeInterval
+import com.cornellappdev.uplift.data.models.TimeInterval
 import com.cornellappdev.uplift.util.PRIMARY_BLACK
 import com.cornellappdev.uplift.util.montserratFamily
