+++ app/src/main/java/com/cornellappdev/uplift/ui/components/gymdetail/GymSwimmingSection.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.cornellappdev.uplift.models.UpliftGym
+import com.cornellappdev.uplift.data.models.UpliftGym
 import com.cornellappdev.uplift.ui.components.gymdetail.TimeFlag
 import com.cornellappdev.uplift.util.ACCENT_CLOSED
 import com.cornellappdev.uplift.util.PRIMARY_BLACK
