+++ app/src/main/java/com/cornellappdev/uplift/ui/components/home/BriefClassInfoCard.kt
@@ -19,8 +19,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import com.cornellappdev.uplift.R
-import com.cornellappdev.uplift.models.UpliftClass
-import com.cornellappdev.uplift.nav.navigateToClass
+import com.cornellappdev.uplift.data.models.UpliftClass
+import com.cornellappdev.uplift.ui.nav.navigateToClass
 import com.cornellappdev.uplift.ui.viewmodels.ClassDetailViewModel
 import com.cornellappdev.uplift.util.GRAY02
 import com.cornellappdev.uplift.util.GRAY03
