+++ app/src/main/java/com/cornellappdev/uplift/ui/components/home/GymCapacity.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.cornellappdev.uplift.models.UpliftCapacity
+import com.cornellappdev.uplift.data.models.UpliftCapacity
 import com.cornellappdev.uplift.util.ACCENT_CLOSED
 import com.cornellappdev.uplift.util.ACCENT_OPEN
 import com.cornellappdev.uplift.util.ACCENT_ORANGE
