+++ app/src/main/java/com/cornellappdev/uplift/ui/components/home/HomeCard.kt
@@ -30,9 +30,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cornellappdev.uplift.R
-import com.cornellappdev.uplift.models.UpliftGym
-import com.cornellappdev.uplift.networking.ApiResponse
-import com.cornellappdev.uplift.networking.CoilRepository
+import com.cornellappdev.uplift.data.models.UpliftGym
+import com.cornellappdev.uplift.data.models.ApiResponse
+import com.cornellappdev.uplift.data.repositories.CoilRepository
 import com.cornellappdev.uplift.ui.components.general.FavoriteButton
 import com.cornellappdev.uplift.util.*
