+++ app/src/main/java/com/cornellappdev/uplift/ui/nav/NavigationFunctions.kt
@@ -1,8 +1,8 @@
-package com.cornellappdev.uplift.nav
+package com.cornellappdev.uplift.ui.nav
 
 import androidx.navigation.NavHostController
-import com.cornellappdev.uplift.models.UpliftClass
-import com.cornellappdev.uplift.models.UpliftGym
+import com.cornellappdev.uplift.data.models.UpliftClass
+import com.cornellappdev.uplift.data.models.UpliftGym
 import com.cornellappdev.uplift.ui.UpliftRootRoute
 import com.cornellappdev.uplift.ui.viewmodels.ClassDetailViewModel
 import com.cornellappdev.uplift.ui.viewmodels.GymDetailViewModel
