+++ app/src/main/java/com/cornellappdev/uplift/ui/screens/ClassDetailScreen.kt
@@ -46,9 +46,9 @@ import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import com.cornellappdev.uplift.R
-import com.cornellappdev.uplift.models.UpliftClass
-import com.cornellappdev.uplift.networking.ApiResponse
-import com.cornellappdev.uplift.networking.CoilRepository
+import com.cornellappdev.uplift.data.models.UpliftClass
+import com.cornellappdev.uplift.data.models.ApiResponse
+import com.cornellappdev.uplift.data.repositories.CoilRepository
 import com.cornellappdev.uplift.ui.components.classdetail.ClassDateAndTime
 import com.cornellappdev.uplift.ui.components.classdetail.ClassDescription
 import com.cornellappdev.uplift.ui.components.classdetail.ClassFunction
