+++ app/src/main/java/com/cornellappdev/uplift/util/Functions.kt
@@ -1,9 +1,9 @@
 package com.cornellappdev.uplift.util
 
 import android.location.Location
-import com.cornellappdev.uplift.models.TimeInterval
-import com.cornellappdev.uplift.models.TimeOfDay
-import com.cornellappdev.uplift.models.UpliftClass
+import com.cornellappdev.uplift.data.models.TimeInterval
+import com.cornellappdev.uplift.data.models.TimeOfDay
+import com.cornellappdev.uplift.data.models.UpliftClass
 import java.util.Calendar
 
 /**
